Pass stage component in event['path']
I'm not clear enough on ASGI to know if this is (1) a bug in Mangum,
(2) a bug in Datasette, or (3) if the spec is open to interpretation such
that both implementations are compliant (yet non-interoperable).

I lean towards perhaps a little of (2) and (3).

For Mangum - in the `handle_http` [1] case, it doesn't use the `stageName` at all.

In the `handle_ws` [2] case, it passes the `stageName` as the
`root_path`, which the WSGI spec defines as:

> The root path this application is mounted at; same as SCRIPT_NAME in
> WSGI. Optional; defaults to "".

For Datasette - it does not appear to use the ASGI concept of root_path
at all.

[1]: https://github.com/erm/mangum/blob/566b0b7569459a8ef6463af08780300521ae61b1/mangum/adapter.py#L87
[2]: https://github.com/erm/mangum/blob/566b0b7569459a8ef6463af08780300521ae61b1/mangum/adapter.py#L132