Add note about wrong absolute URLs
By default, CloudFront doesn't send the original `Host` header to
API Gateway.

This means that API Gateway doesn't send the user-facing `Host` header
to Lambda, which means that Datasette's `absolute_url` method produces
an URL that points directly at API Gateway, which is bad.

We can configure CloudFront to send the original `Host` to API Gateway,
but then API Gateway needs to have an SSL certificate for that host,
which is only possible in the case of a custom CNAME, not in the case of
a generic cloudfront.net domain.