Use CloudFront domain name for absolute URLs
I'm pretty sure this is the "right" way to do this, yet
it has a drawback. :(

This commit adds a Lambda@Edge function that augments the incoming
CloudFront request with an `X-Forwarded-Host` header that contains the
original host name.

We then overwrite the `Host` header that API Gateway provides, so that
Datasette is able to produce valid absolute URLs.

This comes at a minor increase in cost since we're now invoking an
additional Lambda on every request.

Annoyingly, it also means `delete-stack` stops working smoothly. To
cleanly tear down the stack, you have to try a few times. The first
attempt tears down the CloudFront distribution and begins de-replicating
the Lambda@Edge function. It then fails to delete the Lambda@Edge
function, so you need to wait a few hours, and try again. See
https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/lambda-edge-delete-replicas.html
for more info.